fix(fvp): avoid the Scunthorpe problem when looking for "tftf|spm"
A simplistic grep looking for "tftf|spm" will match substrings of the
regular expression (eg "rstspmin" will match). As a result, runs that
are not meant to have the shutdown_on_eot parameter run with it set,
resulting in a hanging successful run since cleanup is skipped.  Tell
grep to only match full words to prevent this.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I663530845b57b1caeb133584e2772d0480a66131