feat(ci): do not apply patches to the main checkout of TF-A
Applying a patch to the main checkout works fine for serial builds (and
CI which is serial) but falls apart when running several builds in
parallel - one job will be building when another tries to apply a patch,
causing a conflict. So make a per-job shallow clone of the repository
whenever a patch needs to be applied.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: Id8099c1515e710db6db7bc904d0d3079e5665b4e