fix(tc4): don't expect the terminal names to be at the start of the line
Just like all FVPs, TC4 doesn't print the terminal name as the first
thing on the line. Update the regex as otherwise the local CI run script
will indefinitely wait for it to match, which it never will.

Since the uarts can be substrings of each other, expect the name to have
a space in front of it as a word delimiter.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I2a8543c8d7a3b5df820e36d8f0da7305aed56d1e