fix(ci): always shut the model down, even if it can exit itself
Some runs can send an EOT character when they are done, and the model
can look for that and turn itself off when it sees it (with
shutdown_on_eot=1). We currently use that to skip the automatic cleanup
(shutting down) of the model. However, that doesn't work well as we
don't have an exact way of specifying which jobs can rely on that
parameter and which can't. Some runs have the parameter but never use
it. Whenever that happens, the cleanup will be skipped and we get a
zombie model, preventing the CI run from finishing (after showing
success).

So never skip the cleanup. Rely on the fact that we will get correct
signals (from expect and the exit status of terminals) to cleanup at the
correct time. Do check that the model exited before we could kill it so
that we don't fail a command (due to set +e) and produce a backtrace.

This is effectively a revert of 359b3ff6c.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I3ec193a6d3d59f818ab8fed8d9b09821076ad0f5