feat: add option to run from local FVP binary
Add option to use local fvp binary with local CI.
Current CI FVP model path tries to use warehouse paths for local CI
run and we cant see a method to override that to a local path rather
than using remote warehouse path.

Add a option to use `retain_paths` and path set in $PATH by user
can be used to run FVP binary needed.

This is all assuming when local CI is run, user is aware of how to
download FVP's from warehouse or other location and use that in $PATH
with `retain_paths`.

Change-Id: I9ef76e097ffe82e5308a8b56b0aa1a6e576216b5
Signed-off-by: Govindraj Raja <govindraj.raja@arm.com>