ci(eclair): scale build jobs with control group CPU quota
We currently scale the number of CPU jobs with the number of CPUs on the
host system. This is fine for a real system, but our Jenkins jobs run
within containers, which are given hard CPU quotas. Consequently, we are
running up against severe CPU throttling during static analysis.

This change ensures that we scale the number of Make jobs with the CPU
quota we have been provided, to mitigate against oversubscription.

Change-Id: I042775cc0f348dec7a2f4ba6e4a2f1f7c8e8232f
Signed-off-by: Chris Kay <chris.kay@arm.com>