fix(tfut): do not bring tfut up when it's not relevant
The run script checks if there's a valid tfut config. If there is, it
will run it. Otherwise it prints that it hasn't been run. Well, this is
not correct, as most configs are not tfut configs and so they will
produce 2 messages - one saying the status of the run config and one
saying that tfut was not run.

Instead, print that tfut was skipped only on the tfut branch of
execution and don't pollute the regular runners with irrelevant
messages.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: Iece8e9708e22678fde34fdb3f5ec7bd0ab42ac91