refactor(ci): print hook logs into the build.log
Hooks build a surprising amount of stuff but their output is not
recorded. This frequently leads to situations where the build failed
after the main `make` command but while still technically building. So
append hooks output to the build.log. Sadly, the hooks can fail the
build so when this happen the caller needs to fail the build too. This
leads to a bit of duplication but is better than a silent exit.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I00c9f38326e27b1783413f5d374d38a54398408f