fix(ci): do not print the build.log on failure
The build output will have already been printed on the console so this
only repeats information that is already there and in somewhat more
unusual failures only hides the real problem with lines upon lines of
output. If the build log is needed for debugging then reading the
build.log is more concise anyway. So drop the print and only leave the
path.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I087ac9d66c1c7460c186f21d0b53f09b6db98370