feat(ci): always print build output to the console
Patch 7c0b09d4 removed printing of the build.log to not obscure any
failures that might happen after the main build. Well, those prints were
generally useful to quickly spot problems and to not have to dig through
build artefacts to figure out what's wrong. So make a compromise -
always print all output to the console and also store the make command's
output in the build.log. This way everything is in one place (the
console) and the build.log is free of rubbish.

Also add a few more places where the build can fail (and therefore
stop) so this is caught a bit better.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I353bf382217237360c0ec10f8bdc6d475349a4a3