fix(ci): don't traverse subdirectories while looking for .test files
Without specifying a max depth, the `find` command will recurse into the
build directories. Directories within them are constantly added and
removed so if we have a few parallel jobs at the same time, a directory
can disappear from under find causing it to fail.

Well the command is only looking for .test files and all of those are in
the top level of the workspace. So don't recurse to avoid the problem
and gain a little bit of efficiency.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I5ea2c1c42a7a6ed545612fe2f4679f5239b31336