refactor(ci): make apply_patch work with dont_clean
dont_clean=1 is really useful for cutting down on build times when we
already have the artefacts present, but the mirror of the repository
poses a problem in that it now has to be kept in sync.

First, don't undo the patches at the end of the build - this helps with
debugging of what the mirror contains. This is a leftover from when
there weren't mirrored checkouts and we had to keep the main checkout
clean. This uncovers a bunch of patch keeping code that's unused and
removed.

Then, always operate on the mirror if applying a patch. It will be
necessary.

Then, clean the mirror and fetch the new diff of the main repository.

Finally, apply the diff and patches as before. This ensures that dirty
builds have the latest diff + patches and not an inconsistent inbetween.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: Ica26e671a445509c743041c481d0f07c4b495f53