ci(coverage): adding shutdown tag to FVP
Providing the capability to shutdown the FVP runs with a user
specified string.  This is used in boot tests that do not terminate
with an EOT write to the uart due to panic or error conditions.  This
would typically be invoked in the run config as:

uart0_shutdown_tag="<termination string>" \

Change-Id: I8407474c0a878d01cbab526d06c81906d7f7224d
Signed-off-by: Mark Dykes <mark.dykes@arm.com>