fix(ci): correctly use the `debug` fragment even with `bin_mode=release`
Configs that specify the `debug` fragment will only pull in debug
binaries. However, they rely on the fact that debug and release binaries
are always built. When running locally with `bin_mode=release` that is
not true and the `bin_mode` argument would take precedence, leading to
every job that relies on being built for debug to fail (largely due to
debug information being suppressed). So make the `debug` fragment
mandate only a build for DEBUG=1 and a corresponding run with it.

Also don't put BIN_MODE in the env as that's not relevant for local CI
and we determine the dynamically and LAVA figures it out separately.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I25db05a0c4472fa6596e59f994a98fd251d7cac0