chore(ci): bump SCP to patched v2.16
As part of pre-release work for TF-A v2.14, update the CI configuration
to use the latest SCP v2.16 patch level.

SCP has moved away from a fixed 6-month release cadence, so while this
appears as a minor version bump, it includes important fixes from the
mainline. The previous update attempt uncovered a bug in SCP, which has
since been resolved upstream.

To help with versioning, also move the hot-patched SCP BL2 binary into
the directory with the other SCP binaries. This should remove any
ambiguity that might arise and having the files strewn in many places.

Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: I08ce7e3d19015b8ff27e99c4fcd283c7a2f2cce7