fix(ci): remove AArch32 etm_trace run configs
They are very old and rely on a plugin of questionable suitability.
With the move away from the warehouse, having plugins in CI will be
harder, so remove the test so we can get rid of the plugin. The aarch64
test remains and works, so the functionality is not completely untested.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I1298559984b1f2e65dbe7d9a9f7b8b9e2a157757