perf(ci): don't run the `memory` command on local runs
We want the output of the `memory` command in upstream CI so we have a
history of what happened to the image size. However, this command takes
quite a long time to run (about a second) and is run once per image. On
local CI runs, where speed is paramount, this is a huge slowdown. Since
there is no history to keep locally, obtaining the output would be
trivial to do manually, and we want rapid development, don't run the
command locally.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: Iffad78537a318370efbbe5afbcbfe6e0ae4fdf36