rfa: work around failing TFTF tests
When running TFTF against RF-A, two tests are failing right now:

 - Test "FF-A callee preserves GP register set per SMCCC"
   in testsuite "FF-A SMCCC compliance".

 - Test "Test ESPI Secure interrupt handling
   in testsuite "FF-A Interrupt".

All other tests are passing. In order to enable TFTF testing in RF-A
CI as soon as possible, this patch temporarily disables these tests
Once the underlying issues are fixed in RF-A, we can revert this
patch.

To disable these tests, we leverage TFTF's PLAT_TESTS_SKIP_LIST build
option and make it point to a text file listing the tests we want to
exclude. As a result, the tests won't be compiled at all into the TFTF
binary.

A new TFTF build config with PLAT_TESTS_SKIP_LIST is introduced for this
purpose and pulled in for the RF-A+TFTF+Hafnium test config.

Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>
Change-Id: I25b7b0f09712f2e4a81abd1be14bff0342b73306